Safety box
The SafetyBox project uses an ESP32 as the main controller to simulate a smart locking system using a single LED. On startup, it checks for saved Wi-Fi credentials in EEPROM — if found, it connects to Wi-Fi and an MQTT broker (HiveMQ) to receive remote commands. If credentials are missing or connection fails, the device switches to Access Point mode, allowing users to enter Wi-Fi details via a configuration web page. Once connected, the ESP32 listens for MQTT commands (“OPEN” or “CLOSE”) to control the LED, representing lock and unlock states. The LED also provides status indication through different blink patterns. The system includes auto-reconnect and returns to configuration mode if the network fails repeatedly.